1570: refactor, fix tests
